Refactor and update various components for compatibility with Minecraft 1.21
- Updated `ItemFrameRendererMixin` to include `partialTick` in `renderNameTag` method.
- Modified `LevelRendererMixin` to adapt to the new `renderSky` method signature and handle PoseStack creation.
- Changed `SoundEngineAccessor` to reflect the new List structure for `tickingSounds` and `queuedTickableSounds`.
- Adjusted `ACBlockRegistry` to use the new method for retrieving nuclear rarity.
- Updated `EntityDropChanceAccessor` to change the parameters for `ac_dropCustomDeathLoot`.
- Enhanced `CommonEvents` to ensure compatibility with the new `ServerLevel` type.
- Refactored `ACItemRegistry` to utilize lazy initialization for custom rarities.
- Updated `ACRarityEnumParams` to use method-based pattern for rarity enum extensions.
- Modified various item classes to use the new rarity getter methods.
- Added `ItemEntityMixin` to implement magnetism features for item entities.
- Created `MagneticDataAccessors` to manage magnetic data for various entity types.
- Updated `enumextensions.json` to reflect changes in rarity parameter methods.
- Cleaned up `neoforge.mods.toml` to remove redundant enum extension declarations.

+++ src/main/java/com/github/alexmodguy/alexscaves/client/event/ClientEvents.java
@@ -339,7 +339,7 @@ public class ClientEvents {
                     ClientProxy.randomTremorOffsets[2] = rng.nextFloat();
                     ClientProxy.lastTremorTick = player.tickCount;
                 }
-                double intensity = tremorAmount * Minecraft.getInstance().options.screenEffectScale().get();
+                float intensity = (float)(tremorAmount * Minecraft.getInstance().options.screenEffectScale().get());
                 ((CameraAccessor) event.getCamera()).invokeMove(
                         ClientProxy.randomTremorOffsets[0] * 0.2F * intensity,
                         ClientProxy.randomTremorOffsets[1] * 0.2F * intensity,
@@ -349,27 +349,27 @@ public class ClientEvents {
         if (player != null && player.isPassenger() && player.getVehicle() instanceof SubmarineEntity
                 && event.getCamera().isDetached()) {
             CameraAccessor cameraAccessor = (CameraAccessor) event.getCamera();
-            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(4.0D), 0, 0);
+            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(4.0F), 0, 0);
         }
         if (player != null && player.isPassenger() && player.getVehicle() instanceof TremorsaurusEntity
                 && event.getCamera().isDetached()) {
             CameraAccessor cameraAccessor = (CameraAccessor) event.getCamera();
-            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(2.0D), 0, 0);
+            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(2.0F), 0, 0);
         }
         if (player != null && player.isPassenger() && player.getVehicle() instanceof AtlatitanEntity
                 && event.getCamera().isDetached()) {
             CameraAccessor cameraAccessor = (CameraAccessor) event.getCamera();
-            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(4.0D), 0, 0);
+            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(4.0F), 0, 0);
         }
         if (player != null && player.isPassenger() && player.getVehicle() instanceof TremorzillaEntity
                 && event.getCamera().isDetached()) {
             CameraAccessor cameraAccessor = (CameraAccessor) event.getCamera();
-            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(10.0D), 0, 0);
+            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(10.0F), 0, 0);
         }
         if (player != null && player.isPassenger() && player.getVehicle() instanceof GumWormSegmentEntity
                 && event.getCamera().isDetached()) {
             CameraAccessor cameraAccessor = (CameraAccessor) event.getCamera();
-            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(12.0D), 0, 0);
+            cameraAccessor.invokeMove(-cameraAccessor.invokeGetMaxZoom(12.0F), 0, 0);
         }
         if (player != null && player instanceof LivingEntity livingEntity
                 && livingEntity.hasEffect(ACEffectRegistry.STUNNED)) {
